changing port numbers datatype to handle only values from range 0-65535

+++ sniffer-mod/src/include/packet.h
@@ -31,9 +31,10 @@ typedef struct {
 typedef TcpPacket *tcpPtr; /* For better readibility */
 typedef UdpPacket *udpPtr; /* For better readibility */
 
-tcpPtr tcp_new(ipv4Ptr src_ip, int src_port, ipv4Ptr dst_ip,
-               int dst_port); /* Create new TCP struct with assigned values */
-void tcp_free(tcpPtr self);   /* Free the memory of struct and values */
+tcpPtr
+tcp_new(ipv4Ptr src_ip, uint16_t src_port, ipv4Ptr dst_ip,
+        uint16_t dst_port); /* Create new TCP struct with assigned values */
+void tcp_free(tcpPtr self); /* Free the memory of struct and values */
 void handle_ip_header(const u_char *ip_header,
                       const u_char *packet); /* Extracts info from ip header */
 void packet_handler(
